+++ src/silo/config/initialize_config.h
@@ -20,7 +20,7 @@ class PreprocessingConfig;
 
 class InitializationFiles {
   public:
-   std::optional<std::filesystem::path> lineage_definitions_file;
+   std::vector<std::string> lineage_definition_files;
    std::optional<std::filesystem::path> phylogenetic_tree_file;
    std::filesystem::path database_config_file;
    std::filesystem::path reference_genome_file;
@@ -30,7 +30,7 @@ class InitializationFiles {
 
    [[nodiscard]] std::filesystem::path getDatabaseConfigFilename() const;
 
-   [[nodiscard]] std::optional<std::filesystem::path> getLineageDefinitionsFilename() const;
+   [[nodiscard]] std::vector<std::filesystem::path> getLineageDefinitionFilenames() const;
 
    [[nodiscard]] std::optional<std::filesystem::path> getPhyloTreeFilename() const;
 
@@ -39,7 +39,7 @@ class InitializationFiles {
    NLOHMANN_DEFINE_TYPE_INTRUSIVE(
       InitializationFiles,
       directory,
-      lineage_definitions_file,
+      lineage_definition_files,
       phylogenetic_tree_file,
       database_config_file,
       reference_genome_file