+++ src/silo/initialize/initializer.h
@@ -16,9 +16,14 @@ class Initializer {
    static silo::schema::DatabaseSchema createSchemaFromConfigFiles(
       config::DatabaseConfig database_config,
       ReferenceGenomes reference_genomes,
-      common::LineageTreeAndIdMap lineage_tree,
+      std::map<std::filesystem::path, common::LineageTreeAndIdMap> lineage_trees,
       common::PhyloTree phylo_tree_file,
       bool without_unaligned_columns
+   );
+
+   static std::optional<common::LineageTreeAndIdMap> findLineageTreeForName(
+      const std::map<std::filesystem::path, common::LineageTreeAndIdMap>& lineage_trees,
+      std::string lineage_tree_name
    );
 };
 }  // namespace silo::initialize